Removed left_child/right_child in favour of a list of children
Changed the Node struct so it no longer stores a reference to a right_child and a left_child. Child nodes are now stored in an array which currently defaults to size 2 to preserve the current functionality.

This sets up the requirement for C4.5 to have an arbitrary number of child nodes when splitting on categorical variables. There are two references to right_child and left_child remaining in the Tree C class which are causing half the tests to fail.